refactor(client): add asyncCreateThunk to userSlice from authSlice, remove authSlice
