+++ packages/client/src/store/index.tsx
@@ -2,14 +2,12 @@ import { configureStore } from '@reduxjs/toolkit'
 import { userReducer } from './userSlice'
 import { playersStatsReducer } from './playersStatsSlice'
 import { gameReducer } from './gameSlice'
-import { authReducer } from './authSlice'
 
 export const store = configureStore({
   reducer: {
     user: userReducer,
     playersStats: playersStatsReducer,
     game: gameReducer,
-    auth: authReducer,
   },
 })
