+++ packages/client/src/store/userSlice.ts
@@ -1,4 +1,11 @@
-import { createSlice } from '@reduxjs/toolkit'
+import { createAsyncThunk, createSlice } from '@reduxjs/toolkit'
+import { useAuth } from '../hooks/useAuth'
+import { IFormData } from '../hooks/useAuth'
+
+export const getAuth = createAsyncThunk(
+  'user/getAuth',
+  async (FormData: IFormData) => useAuth(FormData)
+)
 
 export interface UserState {
   isLoading: boolean
@@ -39,6 +46,34 @@ const userSlice = createSlice({
     toggleAuth(state) {
       state.isAuth = !state.isAuth
     },
+  },
+  extraReducers: builder => {
+    // Add reducers for additional action types here, and handle loading state as needed
+    builder
+      .addCase(getAuth.pending, state => {
+        state.error = null
+        state.isAuth = false
+        state.isLoading = true
+      })
+      .addCase(getAuth.fulfilled, (state, action) => {
+        const { user } = action.payload
+        state.isLoading = false
+        state.isAuth = user.isAuth
+        if (user.isAuth) {
+          state.id = user.id
+          state.firstName = user.firstName
+          state.secondName = user.secondName
+          state.displayName = user.displayName
+          state.login = user.login
+          state.email = user.email
+          state.phone = user.phone
+        }
+      })
+      .addCase(getAuth.rejected, (state, action) => {
+        state.isLoading = false
+        state.isAuth = false
+        state.error = action.payload
+      })
   },
 })
